Renamed plugin SDK utils to plugin_utils to prevent filename conflicts in plugins
